Add files via upload
This project was created with great care and attention as part of our iFood initiative, where we prioritize both user experience (UX) and user interface (UI) design. For the development and implementation of the screens, we utilized HTML, CSS, Bootstrap, and JavaScript to ensure a seamless and efficient user interaction.